Adjust helper text relating to Expo Go
This PR changes the helper text to mention Expo Go to be more precise. Omitted guidance on the managed workflow and prefer not to imply that reflecting native APIs from JS is sanctioned in Expo (of course folks are free to try what they'd like on RN and other platforms -- just prefer different guidance for Expo developers).

Ran Prettier and ESLint to conform with the formatting preferences.